// src/types/socket.types.ts
export interface MessageMedia {
  type: "image" | "file";
  url: string;
  name: string;
  size?: number;
}

export interface MessageReaction {
  emoji: string;
  usersIds: string[];
}

export interface MessageData {
  _id: string;
  conversationId: string;
  senderId: string;
  senderType: "admin" | "user";
  content: string;
  media?: MessageMedia;
  reactions: MessageReaction[];
  readAt?: Date;
  createdAt: Date;
}